requirements.txt
numpy
scipy
pandas
matplotlib
streamlit
scikit-learn
pysoundfile
librosa
pytest
torch==2.8.0
torchaudio==2.8.0
openai-whisper
jiwer
phonemizer
epitran
indic-transliteration
speechbrain
transformers==4.57.6
tokenizers>=0.22,<0.23
huggingface-hub>=0.34,<1.0
accelerate
coqui-tts
